requirements.txt
requests>=2.28.0
aiohttp>=3.8.0
pyyaml>=6.0
tqdm>=4.64.0
rich>=13.0.0
playwright>=1.30.0